Backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
numpy
opencv-python
Pillow
torch
torchvision
tensorflow
gdown